stable-kernel sync: before moving to v3.3-rc1

+++ create_dsp_package.sh
@@ -40,8 +40,10 @@ function libstd_dependicy {
 DIST=$(lsb_release -sc)
 
 if [ $(uname -m) == "x86_64" ] ; then
+ echo ""
  echo "Note: on the x86_64 platform, this script needs (ia32-libs)..."
  echo "--------------------------------------------------------------"
+ echo ""
 fi
 
 cd ${DIR}
@@ -107,6 +109,10 @@ if [ \$(uname -m) == "armv7l" ] ; then
       sudo chmod +x /etc/rcS.d/S61dsp.sh
     else
       #karmic/lucid/maverick/etc
+      sudo update-rc.d -f dsp remove
+      if [ -f /etc/init.d/dsp ] ; then
+        rm -f /etc/init.d/dsp || true
+      fi
       sudo cp /opt/dsp /etc/init.d/dsp
       sudo chmod +x /etc/init.d/dsp
       sudo update-rc.d dsp defaults
@@ -215,7 +221,15 @@ function create_DSP_package {
 	mkdir -p ${DIR}/DSP/lib/dsp
 	mkdir -p ${DIR}/DSP/opt/
 
-	sudo cp -v ${DIR}/dl/TI_DSP_${TI_DSP_BIN}/binaries/* ${DIR}/DSP/lib/dsp
+ if [ -f ${DIR}/dl/TI_DSP_${TI_DSP_BIN}/binaries/baseimage.dof ] ; then
+  sudo cp -v ${DIR}/dl/TI_DSP_${TI_DSP_BIN}/binaries/* ${DIR}/DSP/lib/dsp
+ else
+  echo "--------------------------------------------------------------"
+  echo "Script, failure, DSP bin was not extracted, do you have i32-libs installed?"
+  echo "--------------------------------------------------------------"
+  echo ""
+  exit
+ fi
 
 file-DSP-startup
